Can save and load snapshots by name.

+++ source/botgroup.rb
@@ -14,31 +14,45 @@ class PhlangBotGroup
         @bots.push(bot)
     end
 
-    def save()
+    # Create a snapshot
+    def save(name)
+        # Prepare the data
         data = @bots.map {|b| b.to_h}
         raw = JSON.pretty_generate(data)
 
-        now = Time.now.utc.strftime("%Y-%m-%d@%H:%M:%S")
-        filename = File.join("snapshots", now)
+        # Write the snapshot
+        filename = File.join("snapshots", name)
         File.open(filename, 'w') do |file|
             file.write(raw)
         end
 
+        # Symlink 'latest' to the new file
         latest = File.join("snapshots", "latest")
-
         if File.symlink?(latest)
             File.delete(latest)
         end
         File.symlink(filename, latest)
     end
 
-    def recover()
+    # Recover a snapshot
+    def recover(form="latest")
+        # Work out the filepath
+        filename = File.join("snapshots", form)
+        if File.symlink?(filename)
+            filename = File.readlink(filename)
+        end
+        if !File.exists?(filename)
+            return
+        end
+
+        # Get the raw data
         raw = ""
-        filename = File.readlink(File.join("snapshots", "latest"))
         File.open(filename, 'r') do |file|
             raw = file.read()
         end
         data = JSON.load(raw)
+
+        # Add the data to the group
         data.each do |d|
             @bots.push(PhlangBot.from_h(d, PhlangBotConfig.new(FULL_BUILTINS, MINIMAL_TRIGGERS, MINIMAL_RESPONSES)))
         end
